+++ src/main/java/util/validator/Validator.java
@@ -2,7 +2,7 @@ package util.validator;
 
 import net.itarray.automotion.validation.NewValidator;
 import org.openqa.selenium.WebElement;
-import util.validator.properties.Padding;
+import net.itarray.automotion.validation.properties.Padding;
 
 import java.util.List;
